+++ src/SqlArtisan/Internal/SqlBuilder/DbmsDialect/OracleDialect.cs
@@ -5,4 +5,6 @@ internal sealed class OracleDialect : IDbmsDialect
     public char AliasQuote => '"';
 
     public char ParameterMarker => ':';
+
+    public string CeilFunctionName => Keywords.Ceil;
 }