+++ src/SqlArtisan/Internal/SqlBuilder/DbmsDialect/SqlServerDialect.cs
@@ -5,4 +5,6 @@ internal sealed class SqlServerDialect : IDbmsDialect
     public char AliasQuote => '"';
 
     public char ParameterMarker => '@';
+
+    public string CeilFunctionName => Keywords.Ceiling;
 }