+++ src/SqlArtisan/Internal/SqlPart/Expression/Function/NumericFunction/CeilFunction.cs
@@ -1,0 +1,17 @@
+namespace SqlArtisan.Internal;
+
+public sealed class CeilFunction : SqlExpression
+{
+    private readonly SqlExpression _expr;
+
+    internal CeilFunction(SqlExpression expr)
+    {
+        _expr = expr;
+    }
+
+    internal override void Format(SqlBuildingBuffer buffer) => buffer
+        .AppendCeilFunctionName()
+        .OpenParenthesis()
+        .Append(_expr)
+        .CloseParenthesis();
+}